+++ CHANGES.md
@@ -62,6 +62,7 @@ release:
 
 <!-- Changes to how Black can be configured -->
 
+- Print warning when toml config contains an invalid key (#4165)
 - Fix symlink handling, properly catch and ignore symlinks that point outside of root
   (#4161)
 - Fix cache mtime logic that resulted in false positive cache hits (#4128)